backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.29
pydantic==1.10.12
python-dotenv==1.0.0
PyMySQL==1.1.1
cryptography
langchain==0.1.14
langchain-core==0.1.53
langchain-community==0.0.38
langchain-google-genai==0.0.11
google-generativeai==0.3.2
